slightly refactor helpers to optionally use named exports for repeat and html
