+++ .storybook/helpers/html.js
@@ -1,6 +1,6 @@
 // just using String.raw to tag templates as html literals for linting and highlighting
 // https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Template_literals
-// You may need:
+// You will benefit from:
 //  * Syntax highlighter plugin for your text editor (ex: vscode-lit-html)
-//  * Prettier
+//  * A code formatter like elsint or Prettier that formats on save
 export default String.raw;
