+++ src/components/byline/byline.stories.js
@@ -1,23 +1,21 @@
-import html from "../../../.storybook/helpers/html";
+import { html, repeat } from "../../../.storybook/helpers";
 import { number, withKnobs } from "@storybook/addon-knobs";
 
 export default { title: "Components/Byline", decorators: [withKnobs] };
 
-const authors = () => {
-	return new Array(number("Number of Authors", 2))
-		.fill(authorMarkup())
-		.join(", ");
-};
+const authors = () => repeat(
+	number("Number of Authors", 2),
+	authorMarkup(),
+	", "
+);
 
 const authorMarkup = (author = "Jeanie Choi") =>
 	html`<a href="/">${author}</a>`;
 
-const bylineMarkup = (model) => {
-	return html`<div class="byline">
+const bylineMarkup = (model) => html`<div class="byline">
 		<span class="byline__authors">${model.authors()}</span>
 		<span class="byline__date">${model.date}</span>
 	</div>`;
-};
 
 const data = {
 	authors,
